factor out an IdFilter trait into common module
