+++ kifi-backend/modules/search/app/com/keepit/search/util/IdFilterCompressor.scala
@@ -5,11 +5,13 @@ import org.apache.lucene.store.InputStreamDataInput
 import java.io.ByteArrayOutputStream
 import java.io.ByteArrayInputStream
 import java.util.Arrays
-import javax.xml.bind.DatatypeConverter._
 import java.util.zip.Adler32
 
-object IdFilterCompressor {
-  def toByteArray(ids: Set[Long]): Array[Byte] = {
+object IdFilterCompressor extends IdFilterBase[LongArraySet] {
+
+  override val emptySet: LongArraySet = LongArraySet.empty
+
+  override def toByteArray(ids: Set[Long]): Array[Byte] = {
     val arr = ids.toArray
     Arrays.sort(arr)
     val baos = new ByteArrayOutputStream(arr.length * 4)
@@ -42,7 +44,7 @@ object IdFilterCompressor {
     bytes(4) = (checksum).toByte
   }
 
-  def toSet(bytes: Array[Byte]): LongArraySet = {
+  override def toSet(bytes: Array[Byte]): LongArraySet = {
     val in = new InputStreamDataInput(new ByteArrayInputStream(bytes))
 
     val version = in.readByte().toInt
@@ -75,21 +77,4 @@ object IdFilterCompressor {
     adler32.getValue.toInt
   }
 
-  def fromSetToBase64(ids: Set[Long]): String = printBase64Binary(toByteArray(ids))
-
-  def fromBase64ToSet(base64: String): LongArraySet = {
-    if (base64.length == 0) LongArraySet.empty
-    else {
-      val bytes = try {
-        parseBase64Binary(base64)
-      } catch {
-        case e: Exception => throw new IdFilterCompressorException("failed to decode base64", e)
-      }
-      toSet(bytes)
-    }
-  }
-}
-
-class IdFilterCompressorException(msg: String, exception: Exception) extends Exception(msg) {
-  def this(msg: String) = this(msg, null)
 }
